opsec modules, wininet cradle and some changes to server
Added a cradle that uses Wininet instead of request library.This is useful also for NTLM proxies.
Added opsec pyMM module that uses WinIned instead of requests and imports a minified windows dependency.
It essentially is the same windows dependency but stripped down from unnecessary modules because we use it jsut to access the PEB.
This allows The opsec version to load PythonMM using Wininet and not dropping anything to disk, while the normal pyMM drops windows on disk and does not uses Wininet so there's no NTLM proxy support backed in.
The pyramid server has been also stripped down from the basic authentication because it's meaningless since we are using symmetric encryption and it complicates a bit the usage of WinInet.